+++ MetaTraderStrategies/Classes/CJimBrownTrend.cpp
@@ -9,9 +9,9 @@ public:
     (
         double          inpLots = 1,
         STOPLOSS_RULE   inpStopLossRule = StaticPipsValue,
-        double          inpStopLossPips = 15,
+        int             inpStopLossPips = 15,
         bool            inpUseTakeProfit = true,
-        double          inpTakeProfitPips = 30,
+        int             inpTakeProfitPips = 30,
         STOPLOSS_RULE   inpTrailingStopLossRule = StaticPipsValue,
         int             inpTrailingStopPips = 20,
         bool            inpGoLong = true,
@@ -53,9 +53,9 @@ CJimBrownTrend::~CJimBrownTrend(void)
 int CJimBrownTrend::Init(
     double          inpLots,
     STOPLOSS_RULE   inpStopLossRule,
-    double          inpStopLossPips,
+    int             inpStopLossPips,
     bool            inpUseTakeProfit,
-    double          inpTakeProfitPips,
+    int             inpTakeProfitPips,
     STOPLOSS_RULE   inpTrailingStopLossRule,
     int             inpTrailingStopPips,
     bool            inpGoLong,
